Refactored GNL and printf

+++ refactor/printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: adeimlin <adeimlin@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 16:04:28 by adeimlin          #+#    #+#             */
-/*   Updated: 2025/06/11 17:45:52 by adeimlin         ###   ########.fr       */
+/*   Updated: 2025/06/12 17:42:12 by adeimlin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,5 +17,5 @@
 # include <stdint.h>
 # include <unistd.h>
 
-int		ft_printf(const char *str, ...);
+ssize_t	ft_printf(const char *str, ...);
 #endif
